requirements.txt
lxml
APScheduler
httpx
pyrogram
tgcrypto
telegraph
requests
beautifulsoup4
python-dotenv